requirements.txt
airbyte-cdk>=0.80.0
requests>=2.28.0
psycopg2-binary>=2.9.0
